(fix/refactor): Add validation on user update and password, limited username, password, task description and title
BREAKING CHANGE: In previous versions user presence in DB was checked only for Service create() method. This could have led to major problems with user data collisions. User service update() method has been updated with a check to see if an email or username is being user to prevent this issue.

A password validation mechanism has been added; currently, only the length of the password is checked
Restrictions were set on the length of the username (3 to 32 characters), password length (8 characters or more), task title (up to 510 characters), and task description (2040 characters).